requirements.txt
asgiref==3.5.0
Django==3.1.14
djangorestframework==3.13.1
environs==9.5.0
marshmallow==3.15.0
packaging==21.3
pyparsing==3.0.7
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2